+++ linear_system/matrread.c
@@ -2,7 +2,7 @@
 #include "stdio.h"
 #include "stdlib.h"
 
-void set_b_odd_sum(float* A, float* b, int n) {
+void set_b_odd_sum(double* A, double* b, int n) {
     for (int row = 0; row < n; ++row) {
         b[row] = 0;
         for (int col = 0; col < n; col += 2) {
@@ -12,14 +12,14 @@ void set_b_odd_sum(float* A, float* b, int n) {
 }
 
 
-int read_matr_file(float* A, float* b, int n, char* filename) {
+int read_matr_file(double* A, double* b, int n, char* filename) {
     FILE* file = fopen(filename, "r");
     if (!file) {
         return 1;
     }
 
     for (int i = 0; i < n * n; ++i) {
-        if (fscanf(file, "%f", A + i) != 1) {
+        if (fscanf(file, "%lf", A + i) != 1) {
             return 2;
         }
     }
@@ -29,7 +29,7 @@ int read_matr_file(float* A, float* b, int n, char* filename) {
 }
 
 
-int read_matr_formula(float* A, float* b, int n) {
+int read_matr_formula(double* A, double* b, int n) {
     for (int row = 0; row < n; ++row) {
         for (int col = 0; col < n; ++col) {
             *getel(A, n, row, col) = (rand() % 5000) / 5000.0;
